refactor: sections를 관리하는 일급컬렉션 생성
